- Fixed some include directives - Removed some cast warnings - Integrated MAC Patch from Vurlix (http://www.horde3d.org/forums/viewtopic.php?f=8&t=288)
git-svn-id: https://horde3d.svn.sourceforge.net/svnroot/horde3d/trunk@36 768fa460-af4a-0410-8817-c9c6fcdc2709

+++ Horde3D/Source/ColladaConverter/utils.cpp
@@ -100,10 +100,10 @@ void removeGate( string &s )
 string urlDecode( const string& url )
 {
 	string buffer = "";
-	int len = url.length();
+	size_t len = url.length();
 
-	for (int i = 0; i < len; i++) {
-		int j = i ;
+	for (size_t i = 0; i < len; i++) {
+		size_t j = i ;
 		char ch = url.at(j);
 		if (ch == '%'){
 			char tmpstr[] = "0x0__";
